+++ src/shaders/glsl/line.fragment.glsl
@@ -1,5 +1,4 @@
 uniform lowp float u_device_pixel_ratio;
-uniform bool u_opacity_override;
 
 in vec2 v_width2;
 in vec2 v_normal;
@@ -26,8 +25,7 @@ void main() {
     float blur2 = (blur + 1.0 / u_device_pixel_ratio) * v_gamma_scale;
     float alpha = clamp(min(dist - (v_width2.t - blur2), v_width2.s - dist) / blur2, 0.0, 1.0);
 
-    float finalOpacity = u_opacity_override ? 1.0 : opacity;
-    fragColor = color * (alpha * finalOpacity);
+    fragColor = color * (alpha * opacity);
 
     #ifdef GLOBE
     if (v_depth > 1.0) {