+++ src/shaders/glsl/line_gradient.fragment.glsl
@@ -1,6 +1,5 @@
 uniform lowp float u_device_pixel_ratio;
 uniform sampler2D u_image;
-uniform bool u_opacity_override;
 
 in vec2 v_width2;
 in vec2 v_normal;
@@ -30,8 +29,7 @@ void main() {
     // entire line, the gradient ramp is stored in a texture.
     vec4 color = texture(u_image, v_uv);
 
-    float finalOpacity = u_opacity_override ? 1.0 : opacity;
-    fragColor = color * (alpha * finalOpacity);
+    fragColor = color * (alpha * opacity);
 
     #ifdef GLOBE
     if (v_depth > 1.0) {