+++ src/shaders/glsl/line_pattern.fragment.glsl
@@ -7,7 +7,6 @@ uniform float u_fade;
 uniform mediump vec3 u_scale;
 
 uniform sampler2D u_image;
-uniform bool u_opacity_override;
 
 in vec2 v_normal;
 in vec2 v_width2;
@@ -73,8 +72,7 @@ void main() {
 
     vec4 color = mix(texture(u_image, pos_a), texture(u_image, pos_b), u_fade);
 
-    float finalOpacity = u_opacity_override ? 1.0 : opacity;
-    fragColor = color * alpha * finalOpacity;
+    fragColor = color * alpha * opacity;
 
     #ifdef GLOBE
     if (v_depth > 1.0) {