+++ src/webgl/program/line_program.ts
@@ -17,7 +17,6 @@ export type LineUniformsType = {
     'u_ratio': Uniform1f;
     'u_device_pixel_ratio': Uniform1f;
     'u_units_to_pixels': Uniform2f;
-    'u_opacity_override': Uniform1i;
 };
 
 export type LineGradientUniformsType = {
@@ -27,7 +26,6 @@ export type LineGradientUniformsType = {
     'u_units_to_pixels': Uniform2f;
     'u_image': Uniform1i;
     'u_image_height': Uniform1f;
-    'u_opacity_override': Uniform1i;
 };
 
 export type LinePatternUniformsType = {
@@ -39,7 +37,6 @@ export type LinePatternUniformsType = {
     'u_image': Uniform1i;
     'u_scale': Uniform3f;
     'u_fade': Uniform1f;
-    'u_opacity_override': Uniform1i;
 };
 
 export type LineSDFUniformsType = {
@@ -54,7 +51,6 @@ export type LineSDFUniformsType = {
     'u_mix': Uniform1f;
     'u_lineatlas_width': Uniform1f;
     'u_lineatlas_height': Uniform1f;
-    'u_opacity_override': Uniform1i;
 };
 
 export type LineTextureUniformsType = {
@@ -78,15 +74,13 @@ export type LineGradientSDFUniformsType = {
     'u_mix': Uniform1f;
     'u_lineatlas_width': Uniform1f;
     'u_lineatlas_height': Uniform1f;
-    'u_opacity_override': Uniform1i;
 };
 
 const lineUniforms = (context: Context, locations: UniformLocations): LineUniformsType => ({
     'u_translation': new Uniform2f(context, locations.u_translation),
     'u_ratio': new Uniform1f(context, locations.u_ratio),
     'u_device_pixel_ratio': new Uniform1f(context, locations.u_device_pixel_ratio),
-    'u_units_to_pixels': new Uniform2f(context, locations.u_units_to_pixels),
-    'u_opacity_override': new Uniform1i(context, locations.u_opacity_override)
+    'u_units_to_pixels': new Uniform2f(context, locations.u_units_to_pixels)
 });
 
 const lineGradientUniforms = (context: Context, locations: UniformLocations): LineGradientUniformsType => ({
@@ -95,8 +89,7 @@ const lineGradientUniforms = (context: Context, locations: UniformLocations): Li
     'u_device_pixel_ratio': new Uniform1f(context, locations.u_device_pixel_ratio),
     'u_units_to_pixels': new Uniform2f(context, locations.u_units_to_pixels),
     'u_image': new Uniform1i(context, locations.u_image),
-    'u_image_height': new Uniform1f(context, locations.u_image_height),
-    'u_opacity_override': new Uniform1i(context, locations.u_opacity_override)
+    'u_image_height': new Uniform1f(context, locations.u_image_height)
 });
 
 const linePatternUniforms = (context: Context, locations: UniformLocations): LinePatternUniformsType => ({
@@ -107,8 +100,7 @@ const linePatternUniforms = (context: Context, locations: UniformLocations): Lin
     'u_image': new Uniform1i(context, locations.u_image),
     'u_units_to_pixels': new Uniform2f(context, locations.u_units_to_pixels),
     'u_scale': new Uniform3f(context, locations.u_scale),
-    'u_fade': new Uniform1f(context, locations.u_fade),
-    'u_opacity_override': new Uniform1i(context, locations.u_opacity_override)
+    'u_fade': new Uniform1f(context, locations.u_fade)
 });
 
 const lineSDFUniforms = (context: Context, locations: UniformLocations): LineSDFUniformsType => ({
@@ -122,8 +114,7 @@ const lineSDFUniforms = (context: Context, locations: UniformLocations): LineSDF
     'u_crossfade_from': new Uniform1f(context, locations.u_crossfade_from),
     'u_crossfade_to': new Uniform1f(context, locations.u_crossfade_to),
     'u_lineatlas_width': new Uniform1f(context, locations.u_lineatlas_width),
-    'u_lineatlas_height': new Uniform1f(context, locations.u_lineatlas_height),
-    'u_opacity_override': new Uniform1i(context, locations.u_opacity_override)
+    'u_lineatlas_height': new Uniform1f(context, locations.u_lineatlas_height)
 });
 
 const lineTextureUniforms = (context: Context, locations: UniformLocations): LineTextureUniformsType => ({
@@ -146,8 +137,7 @@ const lineGradientSDFUniforms = (context: Context, locations: UniformLocations):
     'u_image_dash': new Uniform1i(context, locations.u_image_dash),
     'u_mix': new Uniform1f(context, locations.u_mix),
     'u_lineatlas_width': new Uniform1f(context, locations.u_lineatlas_width),
-    'u_lineatlas_height': new Uniform1f(context, locations.u_lineatlas_height),
-    'u_opacity_override': new Uniform1i(context, locations.u_opacity_override)
+    'u_lineatlas_height': new Uniform1f(context, locations.u_lineatlas_height)
 });
 
 const lineUniformValues = (
@@ -155,7 +145,6 @@ const lineUniformValues = (
     tile: Tile,
     layer: LineStyleLayer,
     ratioScale: number,
-    opacityOverride: boolean = false,
 ): UniformValues<LineUniformsType> => {
     const transform = painter.transform;
 
@@ -166,8 +155,7 @@ const lineUniformValues = (
         'u_units_to_pixels': [
             1 / transform.pixelsToGLUnits[0],
             1 / transform.pixelsToGLUnits[1]
-        ],
-        'u_opacity_override': +opacityOverride
+        ]
     };
 };
 
@@ -177,9 +165,8 @@ const lineGradientUniformValues = (
     layer: LineStyleLayer,
     ratioScale: number,
     imageHeight: number,
-    opacityOverride: boolean = false,
 ): UniformValues<LineGradientUniformsType> => {
-    return extend(lineUniformValues(painter, tile, layer, ratioScale, opacityOverride), {
+    return extend(lineUniformValues(painter, tile, layer, ratioScale), {
         'u_image': 0,
         'u_image_height': imageHeight,
     });
@@ -191,7 +178,6 @@ const linePatternUniformValues = (
     layer: LineStyleLayer,
     ratioScale: number,
     crossfade: CrossfadeParameters,
-    opacityOverride: boolean = false,
 ): UniformValues<LinePatternUniformsType> => {
     const transform = painter.transform;
     const tileZoomRatio = calculateTileRatio(tile, transform);
@@ -207,8 +193,7 @@ const linePatternUniformValues = (
         'u_units_to_pixels': [
             1 / transform.pixelsToGLUnits[0],
             1 / transform.pixelsToGLUnits[1]
-        ],
-        'u_opacity_override': +opacityOverride
+        ]
     };
 };
 
@@ -218,12 +203,11 @@ const lineSDFUniformValues = (
     layer: LineStyleLayer,
     ratioScale: number,
     crossfade: CrossfadeParameters,
-    opacityOverride: boolean = false,
 ): UniformValues<LineSDFUniformsType> => {
     const transform = painter.transform;
     const tileRatio = calculateTileRatio(tile, transform);
 
-    return extend(lineUniformValues(painter, tile, layer, ratioScale, opacityOverride), {
+    return extend(lineUniformValues(painter, tile, layer, ratioScale), {
         'u_tileratio': tileRatio,
         'u_crossfade_from': crossfade.fromScale,
         'u_crossfade_to': crossfade.toScale,
@@ -241,12 +225,11 @@ const lineGradientSDFUniformValues = (
     ratioScale: number,
     crossfade: CrossfadeParameters,
     imageHeight: number,
-    opacityOverride: boolean = false,
 ): UniformValues<LineGradientSDFUniformsType> => {
     const transform = painter.transform;
     const tileRatio = calculateTileRatio(tile, transform);
 
-    return extend(lineUniformValues(painter, tile, layer, ratioScale, opacityOverride), {
+    return extend(lineUniformValues(painter, tile, layer, ratioScale), {
         'u_image': 0,
         'u_image_height': imageHeight,
         'u_tileratio': tileRatio,
@@ -261,7 +244,7 @@ const lineGradientSDFUniformValues = (
 
 const lineTextureUniformValues = (
     painter: Painter,
-    layer: LineStyleLayer,
+    opacity: number,
     textureUnit: number
 ): UniformValues<LineTextureUniformsType> => {
     const matrix = mat4.create();
@@ -273,7 +256,7 @@ const lineTextureUniformValues = (
         'u_matrix': matrix,
         'u_world': [gl.drawingBufferWidth, gl.drawingBufferHeight],
         'u_image': textureUnit,
-        'u_opacity': layer.paint.get('line-opacity').constantOr(1)
+        'u_opacity': opacity
     };
 };